training_models/requirements.txt
# YOLO11训练依赖包
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
opencv-python>=4.5.0
Pillow>=8.3.0
PyYAML>=6.0
matplotlib>=3.3.0
seaborn>=0.11.0
pandas>=1.3.0
scipy>=1.7.0
tqdm>=4.62.0
tensorboard>=2.8.0

# 可选：模型导出依赖
onnx>=1.12.0
onnxruntime>=1.12.0

# 可选：TensorRT支持（如果需要）
# nvidia-tensorrt

# 可选：ONNX优化
# onnxsim 